requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
docker==7.1.0
exceptiongroup==1.3.0
google-ai-generativelanguage==0.6.15
google-api-core[grpc]==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
numpy==2.2.6
openai==1.82.1
pandas==2.2.3
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.5
pydantic-core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0